frontend/requirements.txt
pyqt6
pyqtgraph
pyserial
numpy
